Move get_adjacent_provinces location
