+++ diplomacy/persistence/province.py
@@ -184,3 +184,11 @@ class Coast(Location):
 
     def get_adjacent_locations(self) -> set[Location]:
         return self.adjacent_seas.union(self.get_adjacent_coasts())
+
+
+def get_adjacent_provinces(location: Location) -> set[Province]:
+    if isinstance(location, Coast):
+        return location.adjacent_seas | {coast.province for coast in location.get_adjacent_coasts()}
+    if isinstance(location, Province):
+        return location.adjacent
+    raise ValueError(f"Location {location} should be Coast or Province")
